refactor(#792): extract the block registry table into catalog_registry.rs (sweep)
catalog.rs 621 -> 448 (leaves the debt list). The static block-type registry
(block_registry() + BlockRegistryEntry + no_color_override + the fn-pointer type
aliases) moves to catalog_registry.rs as pub(crate); the query functions that
walk it stay in catalog.rs and import it back. No public API change — every
project::catalog::* path is unchanged. project tests (17) pass, zero warnings.
